requirements.txt
streamlit
gigachain==0.2.6 
gigachain_community==0.2.6 
gigachain-cli==0.0.25 
langgraph==0.2.19 
langchain-community==0.2.16
python-dotenv==1.0.1
faiss-cpu
torch==2.2.2
transformers
sentence-transformers